2 functional tests passing, reverted to old testing setup, new way did not work
