+++ project1/tests/conftest.py
@@ -26,18 +26,18 @@ Source: https://testdriven.io/blog/flask-pytest/
 
 '''
 
-@pytest.fixture()
-def app():
-     app = create_app()
-     app.config.update({
-         "TESTING": True,
-     })
+# @pytest.fixture()
+# def app():
+#      app = create_app()
+#      app.config.update({
+#          "TESTING": True,
+#      })
 
-    # other setup can go here
+#     # other setup can go here
 
-     yield app
+#      yield app
 
-    # clean up / reset resources here
+#     # clean up / reset resources here
 
 
 @pytest.fixture()
@@ -84,12 +84,12 @@ def new_instructor():
 
 # functional test setup below
 
-@pytest.fixture(scope='module')
-def test_client():
-    flask_app = create_app('flask_test.cfg')
+# @pytest.fixture(scope='module')
+# def test_client():
+#     flask_app = create_app('flask_test.cfg')
 
-    # Create a test client using the Flask application configured for testing
-    with flask_app.test_client() as testing_client:
-        # Establish an application context
-        with flask_app.app_context():
-            yield testing_client  # this is where the testing happens!
+#     # Create a test client using the Flask application configured for testing
+#     with flask_app.test_client() as testing_client:
+#         # Establish an application context
+#         with flask_app.app_context():
+#             yield testing_client  # this is where the testing happens!
